Add Split.rb
